requirements.txt
flask
mysql-connector-python
passlib
numpy
python-dotenv
